+++ packages/model/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./client.js";
 export * from "./reasoning.js";
 export * from "./sse.js";
+export * from "./stream-recovery.js";
 export * from "./stream-events.js";